+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.24.0] - 2026-09-13
+
+### Fixed
+
+- Fixed the method `CreateProjectBranchFunctionDeployment` to enable Neon functions provisioning.
+
 ## [v0.23.0] - 2026-09-13
 
 The release incorporates the up-to-date [API contract](openAPIDefinition.json) as of 2026-09-13 00:26:00 GMT.